C:/Program Files/Git/

+++ sq_using_binary.cpp
@@ -3,10 +3,10 @@ using namespace std;
 
 int sqroot(int num)
 {
-    int s = 0;
-    int ans = -1;
-    int e = num;
-    int mid  = s + (e-s)/2;
+    long long s = 0;
+    long long ans = -1;
+    long long e = num;
+    long long  mid  = s + (e-s)/2;
     while (s<=e)
     {
         if(mid*mid == num)
@@ -37,7 +37,6 @@ double precion_part(int n,int ans,int n1)
         for(double j = temp; j*j<n1 ; j=j+fraction)
         {
             temp = j;
-            // cout<<j<<" ";10
         }
     }
     return temp;
